requirements.txt
streamlit==1.41.0
google-generativeai==0.8.3

